fix: rename ably api key env var
This is to avoid clashes (Vite does not overwrite existing vars) with other environment variables that a user may have set. It removes a source
of clashes for Ablyians working on it (as ABLY_API_KEY is often used for communicating with the production service).

I would suggest we change this back again in the future once things are stable in production.