messages: make messages an interface and hide underlying eventemitter docs
The messages class was displaying all the internal eventemitter docs and there was no way to prevent it
in typedoc. Unfortunately the no inheritence plugin didn't seem to work.

By making the messages class an interface, we can expose just the information that is usefult to users and
hide internal implementation details from our type documentation.

There's a lot of them and we don't want to advertise to users that they exist, as they should
be internal.