presence: fix flakey integration test
The test could fail because we could subscribe to presence only after the update event had come through.

This change fixes the test.

Also adds the dequal package for deep-object equality to our dev dependencies only. If we need similar functionality
in prod, we can definitely raise the package up (it's very small!).