Make demo app multi-room by accepting a `room` query parameter.
On load the page changes history to the default "abcd" room if no other room is set. If you change the URL you use a different room.

Useul for demos where you don't want the history of your own testing to leak, and also useful if you want to demo to sub-groups of people.