room: improve room lifecycle, handle retries properly
This change makes a number of significant changes to the room lifecycle manager.

Firstly, we make sure that retries are properly handled - including subsequent attachment failures. There are tests
to show multiple failures in a row.

The behaviour of attach is now slightly modified. In the event of an attach failure, we immediately return the result
and update the room status. However, we perform subsequent actions in the background: either rolling back all channels
or waiting for the problem channel to re-attach itself. This is inline with the channels implementation, which will enter
suspended on attach timeout but eventually recover itself. In addition, we no longer consider channels going directly from
attaching to detached, as this is not possible in the spec.

The behaviour of detach remains atomic, all channels must reach some resolution before the operation fails.