test: run react-hooks tests in browsers using playwright
This change introduces browser-based testing for our upcoming react-hooks package. As Vitest
doesn't currently support mocking in the browser, we'll be limited to running integration tests
only.

We add react-hooks test runs in Chromium, Firefox and Webkit.

Also switch the coverage engine from v8 to istanbul, as istanbul
works for testing in browsers.