refactor: convert all typescript files to kebab-case
Convert all the files in the repo to use kebab-case.

We decided to do this for consistency and to have a standard that is enforcable via eslint. It also
saves us from case-sensitivity (or lack thereof) in operating systems causing issues.

To enforce the naming convention, we add eslint-plugin-unicorn and it's filename-case rule.