demo-app: add occupancy to the demo app.
A new OccupancyComponent was created to display the current count of users connected to the room. An occupancy hook was added to listen to the occupancy events and update the counts. The new component has been added to the App.tsx file, so it sits at the same level as the chat message and reaction components. Additionally, the RoomProvider was updated to support occupancy defaults. A new CSS file was also added for stylistic purposes of the OccupancyComponent.