config: unify development under single tsconfig, add module-specific configs for build
In lots of IDEs, the root tsconfig.json needs to be main one - for example, vscode won't find our Jest
extensions because they aren't covered in the base config.

This change makes the root-level tsconfig a global config that covers all code in Chat. That way, our IDEs
can detect all the code and various Jest extensions. In doing so, we move the vite.config.ts and tsconfig.json
for the core project into the core folder and these are referenced only on build.