react-hooks: add useChatClient hook, related context and provider
This commit includes the addition of the `useChatClient` custom hook that allows the user to retrieve the current chat client from the nearest context. A `ChatClientContext` and `ChatClientProvider` are also added, which are necessary for using the new hook. These components ensure that there is only a single instance of the chat context at any given level in the component tree. This commit also includes related tests and adjustments to tsconfig.json configuration to add the new files.
The demo app has also been updated to use the new provider, all usages of the chat client have be replaced with a call to get the chat client from the context.
Added README to the react directory to explain new functionality.
Added new agent for react hooks.