react: add initial tests for room provider and hook
react: add useRoom hook with related context and provider

Adds RoomProvider, RoomContext and useRoom hook to the react hooks and
adapts the demo app to use them.

This version uses two parameters on RoomProvider:

- `attach` controls whether RoomProvider attaches the room on mount and
automatically detaches on unmount. Defaults to true.
- `release` controls whether the room will be automatically released
on unmount. Note that releasing a room also detaches it. Defaults to
true.

The two parameters are useful for cases where the same room must be
provided more than once and a single component's mount/unmount is not
suitable to control attach, detach and release behaviors.

We have a known issue with the message channel auto-attaching on
presence.get, and the typing indicators channel auto-attaching due to
typing indicators working via presence (but on a different channel).

This issue affects the demo app if RoomProvider is used with release=true
and attach=false.

demo: fix: stop showing yourself as currently typing when writing a message

export DefaultRoomOptions as a struct instead of a class so it can be used directly for a default room