cdn: use UMD version for CDN releases, update library name
The library was not working on the CDN when using the ESM version. This is because ably-js does not export
an ESM variant, and thus the import statement from Chat (which marks ably-js as external during build) fails
to resolve the module.

This change fixes the issue by updating the README to point at the UMD version of Chat when using the CDN: which does
work.

We also update the library names in the Vite config, which allows you to use the Chat SDK like so `const chatClient = new AblyChat.ChatClient()` rather
than having to do `new window["ably-chat-js"].ChatClient()`.