react: fix presence data leakage between tests
A number of tests use a room called "room-id". In doing so, it means that we inadvertently leak stateful information
such as the presence set between tests. This leads to tests that are asserting on presence events being flakey, as they
see presence events from other tests.

This change fixes the issue by ensuring that all tests use a unique room id.