react: use stable references to event listeners in hooks
A number of our hooks allow users to pass in their own listeners to be able to monitor events such as new
messages or occupancy updates, even if the hook exposes the latest data.

At present, we use the listener param in a useEffect to apply listener subscriptions. This causes listener cycling if
the hook is called using an inline listener, as the reference changes every time. This behaviour is not desired, as it may
cause the subscriber to miss events.

This change fixes the issue by introducing a new internal hook, useEventListenerRef. This hook creates a stable reference to
the passed in listener between rerenders and discards subsequent values passed in. Therefore we only consider the original listener
passed in on the first render.

We also replace all instances of listener use in the other hooks with calls to this new hook, so that the listener passed in is maintained.