fix/react: dont do getPreviousMessages if hook ummounts
When running React in strict mode, its possible to call getPreviousMessages between the subscription being removed and the setState
taking effect. This causes an error to be thrown as getPreviousMessages requires the subscription to work.

To resolve this, we check if the hook has unmounted before calling getPreviousMessages and return a dummy value if it has.