react: handle unhandled rejections in ChatRoomProvider
The call to `attach` in ChatRoomProvider can reject if the room is released prior to initialisation
(the "Room released before initialization started" error). The catch statement in the Room constructor
doesn't prevent rejection errors propagating in attach (as attach technically returns a new promise that
rejects with the error from the base asyncOpsAfter promise).

This change fixes the issue in React by handling rejections appropriately.