core/messages: prevent unhandled promise rejections if rooms are released before initialization
At the moment, if you create a room, add a messages subscription and then release the room before
initialization has completed, you get an unhandled promise rejection. This is caused by the getPreviousMessages
promise for the subscription being rejected because the room has been released.

This in particular occurs in React strict mode.

To fix the issue, we add a handler to catch global rejections.