react/presence: fix flakey test with usePresence
This change fixes the usePresence integration test. It was flaking because a leave and
detach in quick success (as the test does via unmounting) can lead to duplicate presence events
as realtime will do an implicit leave for any detachments it deems still in presence.

This would lead to the presence events array having 4 events (instead of the expected 3).

The fix is to assert on seeing specific presence events we expect to see, showing that we get the desired
behaviour, but without falling over because of potential race conditions.