rooms: make room getting async
This change makes the rooms.get() function async.

This greatly simplifies the underlying Room logic - we no longer have async initialisation after the instance is returned,
and can do away with all the complexity that brings. Instead, whether we need to delay returning a room because the previous
is being released is now dealt with at the rooms level. The rooms singleton won't return a value until the previous room has
released properly.

The change therefore removes all of the async-related code in the DefaultRoom type and beyond.

There are also changes to React, namely:

- useRoom now returns Room|undefined
- the context value for a Room provider now contains a promise to the room
- fixed a bug by which a room would be immediately invalidated in state in Strict Mode - this is now handled more gracefully
- added new helper hooks to better compartmentalise room resolution and how that relates to React