core/room: add channel manager, fix bug where channel options not applied
This change fixes the bug described in #411 whereby channel options are overwritten by successive uses of the same
channel, rather than merged together.

In doing this, we also fix the bug in relation to newly introduce spec point CHA-RC3. We were previously depending on the
soft-deprecated behaviour in RTS3c which would allow you to update channel options via setOptions.

The change is achieved by adding a channel manager, whose job it is to register all the options for a given channel
prior to the first channels.get() call, and call channels.get with the amalgamated options.

Closes #411