refactor: combine react and core packages into one
This change combines the core and react packages into one - so that one import is required to use the full suite of functionality.

It also updates the build:

dist/core is the core libraries only, which are used for the CDN upload.
dist/chat is the entire library - both react and chat.