Bugfix: Fix race condition in `usePresence` integration test.
A race condition happens when we unmount the `usePresence` hook.
Typically, we expect the unmounting to trigger a leave event, action by the caller through the `leaveClient()` func.
What happens instead is, we unmount the component, which triggers the detach process (as we are unmounting the ChatProvider too).
The `leaveClient()` races against this and loses, so when it finally executes, the channel state is now `detaching` and the function returns an error.

Replaced `TestProvider` with a reusable `Providers` component so the hook and providers can be controlled separately. Updated tests to rerender `Providers` with dynamic children for better control over component lifecycle.