refactor: remove non-working parsing functions
Early on in development we created functons that parse Ably channel messages and turn them into chat messages. A lot has changed
since then, and the functions no longer work if you use them in production.

We want to address encoding/saving chat messages as a general concept at some point in the future, so for now we'll delete the old
functions and resurrect them when the need arises.