Rework channel mode assignment.
If no channel modes are sent on attachment, we get a default set that includes publish/presence etc. What we used to do in chat is rely on this mechanism, assuming that if all options in a room are enabled, then we don't have to send any modes as the defaults are fine.

However, now that we use annotations, and annotation_subscribe is not given as a default mode, we have to reverse this.
We should always send exactly what channel modes we require, so room options should add/remove modes as needed.