feat: manage rooms in react via reference counting
At the moment, users have to provide the attach and release methods to the room provider to get around
the fact that we don't do any reference counting so can't tell when a room is no longer needed. This gets
messy for users as they have to manually manage the order of provider creation and the options.

This change introduces a room reference manager, that tracks the number of references to a room and
calls the attach and release methods accordingly.