Split MessageReactionEventType in two
Summary reaction events and raw reaction events are unrelated, given
that they are emitted through two separate subscription APIs (i.e.
subscribe vs subscribeRaw).

So, make them completely separate in the API. This makes us consistent
with the rest of the SDK, in which an event's data does not vary by
event type.