Make Messages.delete's signature consistent with `update`
Rename the updateParams parameter to be called `details` (of type
OperationDetails), get rid of the DeleteMessageParams type.

It was confusing that we had two types with "Params" in their name, both
of which related to mutation operations, both with a property called
`metadata`, but where this `metadata` property had completely different
meanings (in one case it was the new metadata for the message, and in
the other it was metadata about the mutation operation).