fix/message-reactions: make implementation of promise-returning methods async
Right now the promise-returning methods are not async. This means in the
case of delete - the exception thrown about not specifying a reaction
name would actually _throw an exception_ instead of rejecting a promise.

Users would expect a rejected promise in this case, so make the methods
async to avoid that.